Moved make_lyra_flare_list(), find_lyra_flares(), remove_lyra_artifacts(), get_lytaf_events() from lyra_event_list.py to lyra.py.
Deleted download_lytaf() and old version of get_lytaf_events()
from lyra.py as they are deprecated by new version of get_lytaf_events().

Deleted lyra_event_list.py
